fix : Flame Thrower

+++ Assets/01.Scripts/EntitySystem/EntityStats.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 [Serializable]
 public class EntityStats
 {
-    public int moveSpeed = 1;
-    public int attackSpeed;
+    public float moveSpeed = 1;
+    public float attackSpeed;
     public int attackDamage;
     public int maxHP;
 }
